Further work on token.
Fleshed out the token class a bit more providing wrappers around
scopes and anything that is a number that should actually be a
timestamp. Overrode the `missing_method` call to provide access to
any field in the payload. Updated the middleware to include the
token class as part of the rack environment.