Changed how the app & API are run.
Due to issues of loading different helpers with the same name in
each of the apps had issues when using a single config.ru. To solve
this split them up to be individually runnable applicaitons. While
refactoring also took the opportunity to consolidate the AUTH0_ISSUER
env variable onto the AUTH0_DOMAIN env variable.